+++ Userland/SampleCodeModule/include/syscallsAPI.h
@@ -16,6 +16,6 @@ extern int sys_read_asm(int fd, char *buffer, int size);
 extern int sys_write_asm(int fd, const char *buffer, int size);
 extern int sys_print_asm(int fd, int size, int coor);
 extern int sys_ticker_asm(int fd, int length);
-extern int sys_memcpy_asm(uint32_t * dest, int length);
+extern int sys_memcpy_asm(uint64_t * dest, int length);
 
 #endif
